requirements.txt
fastapi==0.115.12
httpx==0.28.1
pydantic==2.11.5
uvicorn
pytest-asyncio
